Add implementation of bucket_new_custom function
This commit will add the implementation of bucket_new_custom function.
This function creates the bucket_t structure and initialize it using
the array with consecutive buckets sizes given by the user (except the
size of the last bucket). Last bucket has maximum boundary equal to
infinity, so size of this bucket is also an infinity.