Changed custom_bucket_sizes to custom_bucket_boundaries
This commit will change the logic of distribution_new_custom.
So far this function was getting custom sizes of buckets as an
argument. However it would be more readable if instead having
sizes, we would have the boundaries of buckets. That's why now
this function get boundaries of buckets as a parameter.