Little fixes in benchmark and in functions for the distribution_t structure
This commit will add little fixes in the benchmark and in the functions
for the distribution_t structure. In benchmark now we are sure that
compiler won't delete some lines because of the unused return value and
now this commit introduced checking pointers to the time calculating
variables if there are not equal to null and also if distribution_t
structures are not null. Also in distribution_t structure in function
for initializing the exponential distribution this commit introduce the
fix that ensure us that base is greater than 1.